// Server/API/Routes/CreateCharacter/RaceEnum.ts
    export enum RaceEnum {
        HUMAN = 'มนุษย์',
        ELVEN = 'เอลฟ์',
        ORC = 'ออร์ค',
        TRITON = 'ไทรทัน',
        DWARF = 'ดวอร์ฟ',
        HALFLING = 'ฮาล์ฟลิ่ง',
        HALF_ELF = 'ครึ่งเอลฟ์',
        HALF_ORC = 'ครึ่งออร์ค',
        HALF_TRITON = 'ครึ่งไทรทัน',
        DWARFLING = 'ดวอร์ฟลิ่ง',
        ELVON = 'เอลวอน',

        // Monster
        GOBLIN = 'ก๊อบลิน',
        KOBOLD = 'โคโบลด์',
        FELINE = 'เฟลีน',
        CANINE = 'เคนีน',
        AVIAN = 'สัตว์ปีก',
        INSECT = 'แมลง',
        REPTILE = 'สัตว์เลื้อยคลาน',
        AQUATIC = 'สัตว์น้ำ',
        PLANT = 'พืช',
        DRAGON = 'มังกร',
        GIANT = 'ยักษ์',
        ELEMENTAL = 'ธาตุ',
        CELESTIAL = 'เทวทูต',
        CONSTRUCT = 'สิ่งก่อสร้าง',
        OOZE = 'เมือก',
        SPIRIT = 'วิญญาณ',
        MONSTROSITY = 'สัตว์ประหลาด',
        FIEND = 'อสูร',
        DEMON = 'ปีศาจ',
        UNDEAD = 'ผีดิบ',
        TROLL = 'โทรล',

        // 
        UNDEFINED = 'ไม่ระบุ'
    }